ai-analysis-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
python-dotenv==1.0.0
